Differential encoding and decoding is tested and working properly.

+++ utils.h
@@ -29,7 +29,7 @@ typedef struct encodeHeader
 
 	unsigned short channels;
 	unsigned short bitsPerSample; /* In bits */
-	int differenceLength;
+	unsigned long long differenceLength;
 } enc_hdr;
 
 #endif
